// src/models/RemotePointMallOrderHistoryDetail.ts
export type RemotePointMallOrderHistoryDetail = {
    productIdx: string;
    orderIdx: number;
    createdAt: Date;
    imageUrl: string;
    productName: string;
    productDescription: string;
    pointType: string;
    productPrice: number;
    quantity: number;
    totalPoints: number;
    status: string;
};